removed ProviderIcon from app & moved utils out
