+++ web/app.fluidity.money/app/components/index.ts
@@ -1,6 +1,5 @@
 export { default as Table } from "./Table";
 export { default as ProviderCard } from "./ProviderCard";
-export { default as ProviderIcon } from "./ProviderIcon";
 export {
   ToolProvider,
   useToolTip,
