+++ web/app.fluidity.money/app/types/Provider.ts
@@ -1,0 +1,27 @@
+export type Providers =
+  | "Aave"
+  | "Aldrin"
+  | "Circle"
+  | "Compound"
+  | "Dodo"
+  | "Jupiter"
+  | "Lemniscap"
+  | "Maker"
+  | "Multicoin"
+  | "Orca"
+  | "Polygon"
+  | "Saber"
+  | "Solana"
+  | "Solend"
+  | "Uniswap"
+  | "Sushiswap"
+  | "Fluidity"
+  | "Balancer"
+  | "Oneinch"
+  | "Mooniswap"
+  | "Curve"
+  | "Multichain"
+  | "XY Finance"
+  | "Raydium"
+  | "Lifinity"
+  | "Mercurial";
